Add --skip-quant option

+++ extract.py
@@ -26,6 +26,8 @@ parser.add_argument('--stanza-dir', default='', help='Path to Stanza resources d
 parser.add_argument('--lang-code', default='en', help='Language code (default: en)')
 parser.add_argument('--output', default='onnx/', help='Output folder (default: onnx/)')
 parser.add_argument('--text', default='', help="Text to feed the stanza model for extraction")
+parser.add_argument('--skip-quant', type=bool, action="store_true", help="Don't quantize model")
+
 args = parser.parse_args()
 
 if not args.stanza_dir:
@@ -105,13 +107,14 @@ def extract_onnx(self, inputs):
         model_simp, check = simplify(m)
         onnx.save(m, outfile)
 
-        quant_pre_process(outfile, outfile_optim, skip_symbolic_shape=True)
-        quantized_model = quantize_dynamic(outfile_optim, outfile_quant, weight_type=QuantType.QUInt8)
-        if os.path.isfile(outfile_quant):
-            os.unlink(outfile)
-            os.rename(outfile_quant, outfile)
-        if os.path.isfile(outfile_optim):
-            os.unlink(outfile_optim)
+        if not args.skip_quant:
+            quant_pre_process(outfile, outfile_optim, skip_symbolic_shape=True)
+            quantized_model = quantize_dynamic(outfile_optim, outfile_quant, weight_type=QuantType.QUInt8)
+            if os.path.isfile(outfile_quant):
+                os.unlink(outfile)
+                os.rename(outfile_quant, outfile)
+            if os.path.isfile(outfile_optim):
+                os.unlink(outfile_optim)
     else:
         print("Something went wrong")
         exit(1)
